Alocação de itens dicotômicos nos níveis da escala

Alocar itens dicotômicos em pontos ou níveis da escala

aloca.dico(
  par,
  met,
  dp = 1,
  int.dp = 0.5,
  int.nivel = c(-2, 2),
  centro = "centro",
  banco,
  min.casos = 50,
  log = TRUE
)

Arguments

par

Objeto com os parâmetros dos itens estimados. Deve possuir três colunas, uma para cada parâmetro (nesta ordem: a, b, c). Só é necessário quando o método for um dos que seguem: b, b65, n65 ou (2+c)/3.

met

O método de alocação. As opções são: b (o parâmetro 'b' é utilizado para alocar o item no nível); b65 (o valor de theta em que a probabilidade de acertar o item vale 0.65); b65emp (o valor de theta em que 65 porcento das pessoas acertaram o item; é a abcissa da intercessão da reta 0.65, com a interpolação linear entre a proporção empírica do nível e a do nível anterior); n65 (o nível para b65); n65emp (o nível em que 65 porcento das pessoas acertaram o item); (2+c)/3 (o valor de theta em que a probabilidade de acertar o item vale (2+c)/3)

dp

Desvio padrão da escala transformada

int.dp

Os níveis variam de quanto em quanto, em relação ao dp?

int.nivel

O nível mais baixo e o nível mais alto

centro

'centro' se o nível for centrado no ponto; 'esquerda' se o nível inicia no ponto (fechado na esquerda)

banco

Objeto do tipo data.frame com as respostas (0 = erro; 1 = acerto) e o escore de cada indivíduo. A última variável do banco é oe score. Esse objeto é necessário somente se met = 'b65emp' ou met = 'n65emp'. ATENÇÃO: os nomes das colunas dos itens devem começar com 'I' e o da nota deve ser 'NOTA'; para detalhes, veja o exemplo.

min.casos

Número mínimo de respostas em um determinado nível para que o item possa ser alocado nele. Usado somente quando o método for 'b65emp' ou 'n65emp'

log

Argumento lóico. A métrica é logística?

Value

Details

Para esta função, ainda é necessário pensar sobre as transformações dos parâmetros para alocação dos itens na escala. Por exemplo, se de fato vai para escala (0,1) adequadamente e depois para a escala da avaliação. É possível que haja algum erro nessas transformações.

Author

Alexandre Jaloto

Examples

set.seed(1000)
par = as.matrix(data.frame(a = rlnorm (30, 0, .5), b = rnorm (30, 0, 1), c = rbeta (30, 5, 17)))
pop = rnorm (5000, 0, 1)

# precisa ser do tipo data.frame (não pode ser data.table nem matrix)
banco = data.frame(simular (pop, par))

calib = mirt::mirt(banco, 1, itemtype = '3PL', TOL = 0.01)
#> Iteration: 1, Log-Lik: -89871.287, Max-Change: 0.90545Iteration: 2, Log-Lik: -88841.593, Max-Change: 0.48294Iteration: 3, Log-Lik: -88627.726, Max-Change: 0.29078Iteration: 4, Log-Lik: -88557.612, Max-Change: 0.33090Iteration: 5, Log-Lik: -88520.908, Max-Change: 0.50907Iteration: 6, Log-Lik: -88499.604, Max-Change: 0.25420Iteration: 7, Log-Lik: -88486.429, Max-Change: 0.30219Iteration: 8, Log-Lik: -88478.771, Max-Change: 0.34916Iteration: 9, Log-Lik: -88473.924, Max-Change: 0.37221Iteration: 10, Log-Lik: -88470.845, Max-Change: 0.27773Iteration: 11, Log-Lik: -88468.844, Max-Change: 0.16897Iteration: 12, Log-Lik: -88467.564, Max-Change: 0.03431Iteration: 13, Log-Lik: -88467.369, Max-Change: 0.03340Iteration: 14, Log-Lik: -88466.784, Max-Change: 0.21750Iteration: 15, Log-Lik: -88466.272, Max-Change: 0.01221Iteration: 16, Log-Lik: -88466.261, Max-Change: 0.01212Iteration: 17, Log-Lik: -88466.063, Max-Change: 0.01031Iteration: 18, Log-Lik: -88465.937, Max-Change: 0.00891
nota = data.frame(mirt::fscores (calib))
names (nota) = 'NOTA'
par = mirt::coef(calib, simplify = TRUE, IRTpars = TRUE)$items[,1:3]
banco$NOTA = nota

# atenção para a forma de nomear
names (banco) = c(paste0('I', 1:30), 'NOTA')
head (banco)
#>   I1 I2 I3 I4 I5 I6 I7 I8 I9 I10 I11 I12 I13 I14 I15 I16 I17 I18 I19 I20 I21
#> 1  1  1  1  1  1  1  1  1  1   1   1   0   1   1   1   0   0   1   0   1   0
#> 2  1  1  1  1  1  0  0  0  1   1   1   1   1   1   1   1   1   1   1   1   0
#> 3  1  0  1  1  1  0  1  0  1   1   1   0   1   1   0   0   1   1   1   1   1
#> 4  1  0  1  1  1  1  1  1  1   1   1   0   1   1   1   1   1   1   1   1   1
#> 5  0  0  1  0  1  0  1  0  1   1   1   0   1   1   1   1   1   0   1   0   1
#> 6  1  0  0  1  0  0  0  1  1   1   1   0   1   1   1   0   1   0   1   1   1
#>   I22 I23 I24 I25 I26 I27 I28 I29 I30       NOTA
#> 1   1   0   1   0   1   1   1   0   1  0.3487037
#> 2   1   1   1   1   0   1   1   0   1  0.4460082
#> 3   1   1   0   0   0   0   1   0   1  0.3759419
#> 4   1   1   1   1   1   1   1   1   1  1.8992345
#> 5   1   0   1   0   0   1   1   0   0 -0.6241491
#> 6   1   0   1   0   0   1   1   1   1  0.1758800

int.nivel = c(-3, 3)

centro = 'esquerda'
met = 'n65emp'

aloca.dico(par = par, met = met, int.nivel = int.nivel, centro = centro, banco = banco, min.casos = 30)
#>  [1] -0.5  1.0 -1.0 -0.5 -1.0  0.5 -2.0  1.0 -0.5 -0.5 -0.5  1.0  0.5  0.0  0.0
#> [16]  1.0  0.0  0.0 -1.5 -1.0  0.0 -1.0  0.5 -1.0  0.5  1.0 -1.5  0.5  2.0 -1.0