Alocação dos indivíduos nos níveis da escala

Alocar os indivíduos, de acordo com seu escore, nos níveis da escala

aloca.pessoa(escore, niv = seq(-2, 2, 0.25), direita = TRUE)

Arguments

escore

Vetor com os escores dos indivíduos

niv

Os níveis da escala

direita

argumento lógico. Se TRUE (padrão), é fechado na direita e aberto na esquerda. Se FALSE, é aberto na direita e fechado na esquerda.

Value

A função retorna um objeto do tipo lista com dois elementos:

$theta: O escore dos indivíduos

$niveis: O nível de cada indivíduo na escala

Details

O nível é selecionado de acordo com o escore. O participante é posicionado no nível em que o escore é maior do que o valor mínimo do intervalo e menor ou igual ao maior valor do intervalo, se direita = TRUE Se direita = FALSE, o participante é posicionado no nível em que o escore é maior ou igual ao valor mínimo do intervalo e menor do que o maior valor do intervalo.

Author

Alexandre Jaloto

Examples

set.seed (12345)
escore = rnorm (100)
aloca.pessoa (escore)
#> Error in aloca.pessoa(escore): object 'met' not found